Enhancements to the tokenizer (another version).
Main changes is to remove the FSM, and separate the input from accumulator, and
therefore, provide separate functions to handle input and accumulator.

Another enhancement is the use of "ecase" for atom-handling, instead of
"case", to avoid the need for a catch-all case. Also note that multiple
values can be passed in a list, if they are handled similarly.

Future changes (maybe):
- Back to passing functions, rather than relying on generics.
- May remove the "tokenize" function, since "create-tokenizer" is
simpler, with no loss of functionality.
- Move test code, and use testing framework (five-am?)