Refactor regex by making NFA construction direct.
Rather than transforming input regex into a tree of element objects,
and then generate NFA for that tree of objects, I created a generic
function (REGEX-TO-NFA) to do the NFA generation directly.

The benefits:
- No need for most NFA elements (removed in this revision).
- Now the code in PREPARE-REGEX-TREE and PRODUCE-NFA is combined into the
  new generic function (REGEX-TO-NFA), and it becomes simpler.

Also in this commit: I made some clean-up (slots and methods that were
marked for removal).