Advance input by 1 char, in case of match failure to allow for retry.
* Main Change:

Keep track of position of matching operation start. It is used upon
matching termination, to backtrack to this position + 1, for the next
matching operation. This is the new behavior of the matching operation,
which allows it not to give up in case of no match, but to advance in the
input char by char and try to match again. This behavior is implemented in
the NOTIFY-MATCH-TERMINATION method, which means that currently, it's left
up to the input source implementor (I'm still not decided on this, but
committing now since it's currently working (no extensive tests done yet
though).


*Other Changes:

- Add generic method for input, returning the remaining length (number of
  characters not yet read). It is currently used only in testing code, but
  could be useful in general.
- Since now we advance the input even in case no characters are
  accumulated, I changed the condition of terminating the test loop to
  "input empty" (rather than "nothing is accumulated"), without risking
  infinite loop.
- Some comments and TODO lines added to source code (more
  cleanup/refactoring to come in next commits).
- Extra closing parentheses removed.