Remove boundary check when advancing input reading position.
Rationale: in normal flow, where the input is handled under control of
the matching function, we never get beyond boundary, since INPUT-EMPTY-P
is used before advancing. So this check is redundant, and it could also
mask other bugs. So I'm removing it for both efficiency reason, and also
to catch any bug that it could be masking.

Another option is to throw an error in case of advancing beyond boundary
(i.e. putting the check back), but not sure at this point if it would be
useful.