Enhancements to the input source, mainly: consumption VS accumulation.
This is an important commit, carrying a main change to the input source
handling. It has no impact on the regex matching function though (which
probably indicates that separating the input source logic was a good idea).

We start to distinguish between accumulated value and consumed value:
- accumulated value is part of the text that was lastly matched.
- consumed value is part of the text that was consumed, whether in a
  match success or failure operation.

So a new generic method is added to the input source interface to retrieve
the last consumed value.

Following changes added to the basic-regex-input implementation:

- New slot for accumulator-end, which now together with accumulator-start
  will determine the last accumulated text (in matching success).
- Defined two flags to decide whether to advance input position (cursor)
  in case of no consumption. I think it's good that the MATCH-REGEX
  function is not concerned with this functionality, which may be omitted
  in another (simpler) input implementation altogether.
- The NOTIFY-MATCH-TERMINATION method is almost rewritten to support these
  changes.
- Another unrelated change is the introduction of REMAINING-LENGTH slot,
  which is currently used in testing/debugging.

Testing code is also updated/enhanced accordingly, and now we have
different cases of backtracking handled, and also the distinction between
consumption and accumulation is demonstrated, including the tricky case
of successful match without accumulation (e.g. matching (* X) against "A",
which succeeds (since zero-or-more), but does NOT accumulate the 'A" (it
is only consumed).