Introduce package nicknames instead of USING the packages.
For better readability and name space segregation, I'm adding local
nicknames for most packages, and removing those packages from the list of
used packages. So from now on, my packages need to be explicitly prepended
to symbols in the client packages. I'm keeping USE for CL and other
well-known packages.

Note: I chose local nicknames to give control to the client code to
      choose the nickname, and also to make it clear in the client package
      definition which packages it references.