Refactoring of FSM-TRAVERSAL into FSM, acceptance predicate method added.
Removed the FSM-TRAVERSAL package, replacing it with an FSM package, to
contain features that are common to NFA and DFA. Now it defines two
generic methods:

- FSM traversal (moved from FSM-TRAVERSAL)
- FSM acceptance state predicate (newly added)

Implementations for the above are now in the NFA and DFA packages (FSM
traversal implementations moved from FSM-TRAVERSAL).

TODO: DFA is still in REGEX package, and should be moved to a dedicated
package in future refactoring.