Graphviz export: add ability to use state object address as node label.
Add keyword argument to allow using the state object's address as node
label in the diagram. I added this hoping that it could help in debugging
(to be able to correlate trace with diagram), but still not perfect as
trace displays addresses in hexadecimal, while I use the address in
decimal format (may adjust later).

Note: retrieving object address is implementation-specific, and currently
only SBCL is supported. The fallback is to index-based labels.