DFA updates, in preparation for regex negation (coming in later revision).
The change to dfa-state-definitely-terminal-p is explained in the
documentation string, and still under evaluation.

The added function dfa-state-dead-end-p may not be needed, kept for now as
it might be used in analysis/testing, not necessarily in functional code.

Note: based on ongoing analysis, I'll either keep or revert this commit.