Regex negation partial implementation (1st iteration).
Changes:
-------
1. Update zero-or-one element by introducing extra auto-transitions, in
order to comply to Thompson's construct. I needed to do this in order to
split the entry state into two states, one of them is connected to output,
and the other is not. Without it, the current negation would fail.
2. Update or-element in a similar way, although I did not verify yet
whether this is needed or not.
3. Also added extra auto-transition to the one-or-more element, to comply
to Thompson, although not sure if this is needed.
4. Current negation is implemented as explained below.
5. Commenting out negated-nfa-state-p.

Negation Details:
-----------------
Current negation implementation is based on identifying NFA states of the
element to be negated in one of 3 categories:
a) Inner's output state (we call it "continuation point" or CT), or a
state connected to the CT via a series of auto-transitions.
b) Dead-end state, which is a state that is not connected to CT, and having
normal transition(s).
c) Absolute dead-end state, which is a state that is not connected to CT,
and having no normal transitions. This normally results from an inner
negation. Example showing its significance: (not (not a)).
normal transition(s).

After identification, negation is handled as follows:
- New output state created for the negation element
- No change to CT and CCT states (will become dead-ends)
- Inner dead-ends are connected to output state via a new (+ any-char)
element.
- Inner absolute dead-ends are connected to output state via an
auto-transition.

There are still problems to solve in this implementation, for example,
(not (not <regex>)) does not give accurate NFA, but I'm committing now in
order to compare results with upcoming changes and detect any regressions.