Regex negation: 2nd iteration (partial).
More changes related to regex negation, w.r.t. 1st iteration (2c3352c).

Details:
-------
1. Add dead-end flag to NFA state, to be set/unset upon negation, depending
on the state's type (dead-end/continuation point).
2. Change how dead-end states are connected to output state, based on
whether the state is in a CT/CCT closure or not.

In a future iteration, the dead-end flag will be checked during matching
to stop the matching and report matching failure. Example showing the
effect:
- consider regex (not (* "ab")).
- NFA generated for this regex would match text such as "abc", rather than
rejecting it once "ab" is detected. This is because technically, "abc" is
in the set of all texts, excluding {"", "ab", "abab", ...}, so normally it
should match, even though this is probably not the expected result by the
user.
- The dead-end flag will allow the matcher to change this behavior and stop
after "ab", indicating matching failure.
- I will also add another flag to control the matcher's behavior, by
enabling/disabling this check.

TODO: better (more efficient) implementation for the
inner-continuation-point-closures list preparation (we need to include
only the dead-end states, not all states of the closure).
TODO: greedy/non-greedy implementation.
TODO: correct implementation for the any-other-char transitions, since
currently it won't work correctly with multiple negations.