Regex negation: 7th iteration (dead-end paths clean-up).
So far, handling of transitions on any-other-char upon negation depended
on the type of the origin state (rev 84d78dd). In this commit, we change
the handling to be dependent on the type of destination state, as follows:
Since we add transitions on any-other-char to continuation point during
negation, then when negating again (an outer negation), we check if this
transition is going to CT. If so, then this CT will become a dead-end
upon negation. Then, we can delete this any-other-char transition
altogether.

I also added another clean-up for all auto transitions that lead to
the inner continuation point, since with negation, these will lead to a
dead-end, and hence will be dead paths.