Regex negation: fix regression in handling of any-other-char transition.
Iteration #8 (rev 2224924) removed the logic of toggling the transition on
any-other-char, based on which the decision of adding auto-transition to
new CT was made. So before this commit, the auto-transition may be added
in some cases where the transition on any other is removed.

I fixed the problem by changing the returned value of setting the
transition on any-other-char, to be NIL if the transition is already set
(rather than returning the currently set value).