Update regex sexp parsing to allow reading element tags in any package.
With this change, regex tags (SEQ, CHAR-RANGE, OR etc.) do not have to be
keywords any more, so a regex could be defined like this:

(seq #\a (char-range #\x #\z))

instead of this:

(:seq #\a (:char-range #\x #\z))

The reason is to make typing regex a little bit easier. Note however that
using keywords is still also allowed (since I'm now allowing to read those
symbols in any package). That's why I don't have to update the unit tests
in this revision.

I also defined a macro, to avoid repeating the prefix of the package name
in each ECASE line.

I may review/simplify in later revisions.