Partial implementation of greedy/non-greedy negation.
I added a flag to the negation regex to indicate the greedy/non-greedy case.
It is NOT implemented on the user interface side (sexp), but only in the
regex itself, and is initialized to non-greedy. So effectively, there is no
way (at least for now) for the user to use the greedy feature.

Based on this flag, during NFA of negation, we either add a (* any-char) before
the continuation point or not. And since it is initialized to "non-greedy",
we won't generate such element from this point on.

Rationale: It simplifies the generated NFA and DFA (for my research and
testing), and more importantly, I think the greedy use case is confusing, and
having no important value (to be reconsidered later).