Add class-util package.
Add a class-related util package. It currently provides macros that simplify
creating a class whose slots are all functions (typically closures). The
provided macros are for creating the class definition, as well as to create a
construction macro.

I may use these in upcoming commits (alternative regex-input implementation).

TODO: list-of-symbols type is defined here for now. May move it later to the
symbols util (or elsewhere?).