Add class util macro to simplify calling function slots.
Add macro with-function-slots-funcall-macros that defines local macros to
simplify FUNCALLing function slots, without introducing runtime overhead.
This macro uses another macro let-slots, which is useful in general for
read-only slots, that we would want to read once, and use multiple times (e.g.
in a loop), without accessing the slot each time.

The let-slots macro is generic enough to also be useful on its own, hence,
it's also exported.

The with-function-slots-funcall-macros will be used in upcoming commit that
enhances the regex input interface.