Refactoring: use a new NFA traversal abstraction in inversion part of regex negation.
Introduce `with-unique-visit` macro for recursive traversal (of a graph). It abstracts
away the hash table lookup and the local function creation. This removes boilerplate
code, and reduces nesting (and hence indentation).

This macro is used in the element inversion part of the negation function. Also I made
use of the freed horizontal space to enhance code formatting.