Experimental (and not fully working): make set of acceptance of negation more permissive.
This change will most probably be reverted. I just wanted to record the experiment. To
understand the change, look at the diff of the test cases. Note that test 10 does not
pass. But I'm stopping this experiment here as I'm revising the concept of negation. I
think I'll settle at the idea of aborting regex matching, as long as the negated regex
is matched (i.e. once the inner continuation point is reached). So for example, the
text "ABA" would NOT match the regex (* "AB"), once "AB" is consumed. I'll handle this
after I revert this one (check upcoming commits).