Fix bug in negation: exclude adding transition on any-other to inner dead-ends.
When an inner element-connected state is dead-end (i.e. inner element is another
negation), should not add transition on any-other towards new continuation point.
This was discovered with regex (not (not (+ "AB")), which is now added as a test case.

Three new test cases added, including two for one-or-more element.