Refactoring: rename DFA's "candidate terminal" into "candidate matching point".
The intended meaning of this flag is matching, not just termination. I updated
documentation strings to indicate that some issues are open regarding the DFA state
type (matching/dead-end), also the concept of dead-end in NFA VS DFA. I'm investigating
these issues in the context of regex negation (in progress).