Insert state in confirmed-states-table as soon as it's confirmed within the loop.
Rather than waiting for the loop (normal/any-char/auto) to finish, insert confirmed
states as soon as they are found to be confirmed, for the update to be available within
the recursion. This has no effect, at least for current test cases, and I didn't think
about possible NFAs that would profit from this change, but this change is reasonable,
since the earlier the status is updated and visible during recursion, the better.

Other minor form relocation is included.