Replace separate confirmed states table with simpler and more efficient solution.
Rather than keeping track of the confirmed (resolved) states in a separate table, use
the pending states table, with different value to indicate pending/resolution (:pending/
:resolved). And now in order to detect whether a new loop iteration is needed, we use
flags to be raised when there is need for more iterations, that is, when more pending
state(s) is/are newly discovered.

The pending-states-table is now renamed to resolution-progress-table.

TODO: consider omitting even this one, and keeping all info in the output.