Change criterion for accepting DFA state to take into account negation's dead-end status.
I was thinking about the interaction between terminus and dead-end statuses for a while,
and in this revision, I added a test case that demonstrates the issue: when a DFA state
is terminus (has terminus NFA state in closure), it is considered accepting, however,
if a dead-end NFA state is also in the closure, the accepting status should be cancelled.
I did this by adding `acceptance-nfa-closure-union-p` and using it instead of
`terminal-nfa-closure-union-p`.

NOTE (TODO): this is not 100% settled, as I may revise the rules after negation is
finalized (handling tolerance/intolerance, greedy/non-greedy). More test cases still to
be added. But for now, all test cases pass.