Refactoring: create and use `make-char-range-splitting-points-extractor` factory.
The splitting points collection functionality actually belongs to the elements package,
not the NFA states package. Besides, it will also be used in the inversion regex
handling, where the input will be a vector of elements (not a list of NFA states). For
this reason, the main code is extracted and provided through a factory interface, that
will be reused in all uses cases involving collection of splitting points.

This commit includes:
- add factory function
- use factory function to modify collection function with input NFA states (client code
  won't need to be changed since function keeps its signature).
- use factory function to define collection function with input elements (will be used by
  inversion in upcoming commit).

Note: both collection functions are having the same name, but they are defined in
two different packages.