Update regex inversion implementation to use element inversion (like in negation).
The updated implementation is the more accurate one, since this one does not rely on an
"any-other" element (which lacks context, as was faced and updated in negation).
The added test cases 5 and 6 pass with this implementation (wouldn't pass with the
previous one).

Also two test cases are added for the handling of redundant elements in inversion, and
one test case added for element inversion (also for handling of redundant elements).